+++ src/Facades/Auditor.php
@@ -3,6 +3,7 @@
 namespace Butler\Audit\Facades;
 
 use Butler\Audit\Auditor as AuditorClass;
+use Butler\Audit\Testing\AuditorFake;
 use Illuminate\Support\Facades\Facade;
 
 /**
@@ -12,6 +13,9 @@ use Illuminate\Support\Facades\Facade;
  * @method static \Butler\Audit\Auditor initiator(string $initiator, array $context = [])
  * @method static \Butler\Audit\Auditor initiatorContext(string $key, mixed $value)
  * @method static void log(string $event, array $eventContext = [])
+ * @method static void assertLogged(string $eventName, \Closure $callback = null)
+ * @method static void assertNotLogged(string $eventName, \Closure $callback = null)
+ * @method static void assertNothingLogged()
  *
  * @see \Butler\Audit\Auditor
  */
@@ -20,5 +24,12 @@ class Auditor extends Facade
     protected static function getFacadeAccessor()
     {
         return AuditorClass::class;
+    }
+
+    public static function fake(): AuditorFake
+    {
+        static::swap($fake = new AuditorFake('uuid'));
+
+        return $fake;
     }
 }